refactor: DefaultTracingOptions into common
