+++ google/cloud/internal/populate_common_options_test.cc
@@ -212,6 +212,28 @@ TEST(DefaultTracingComponents, WithValue) {
   EXPECT_THAT(actual, ElementsAre("a", "b", "c"));
 }
 
+TEST(DefaultTracingOptions, NoEnvironment) {
+  ScopedEnvironment env("GOOGLE_CLOUD_CPP_TRACING_OPTIONS", absl::nullopt);
+  auto const actual = DefaultTracingOptions();
+  auto const expected = TracingOptions{};
+  EXPECT_EQ(expected.single_line_mode(), actual.single_line_mode());
+  EXPECT_EQ(expected.use_short_repeated_primitives(),
+            actual.use_short_repeated_primitives());
+  EXPECT_EQ(expected.truncate_string_field_longer_than(),
+            actual.truncate_string_field_longer_than());
+}
+
+TEST(DefaultTracingOptions, WithValue) {
+  ScopedEnvironment env("GOOGLE_CLOUD_CPP_TRACING_OPTIONS",
+                        "single_line_mode=on"
+                        ",use_short_repeated_primitives=ON"
+                        ",truncate_string_field_longer_than=42");
+  auto const actual = DefaultTracingOptions();
+  EXPECT_TRUE(actual.single_line_mode());
+  EXPECT_TRUE(actual.use_short_repeated_primitives());
+  EXPECT_EQ(42, actual.truncate_string_field_longer_than());
+}
+
 TEST(MakeAuthOptions, WithoutTracing) {
   auto options = Options{}.set<EndpointOption>("endpoint_option");
   auto auth_options = MakeAuthOptions(options);
