+++ google/cloud/internal/populate_grpc_options_test.cc
@@ -15,6 +15,7 @@
 #include "google/cloud/internal/populate_grpc_options.h"
 #include "google/cloud/credentials.h"
 #include "google/cloud/grpc_options.h"
+#include "google/cloud/internal/populate_common_options.h"
 #include "google/cloud/testing_util/scoped_environment.h"
 #include "absl/types/optional.h"
 #include <gmock/gmock.h>
@@ -61,26 +62,12 @@ TEST(PopulateGrpcOptions, WithSetEmulator) {
   EXPECT_NE(actual.get<GrpcCredentialOption>(), credentials);
 }
 
-TEST(PopulateGrpcOptions, DefaultTracingOptionsNoEnvironment) {
-  ScopedEnvironment env("GOOGLE_CLOUD_CPP_TRACING_OPTIONS", absl::nullopt);
-  auto const actual = internal::DefaultTracingOptions();
-  auto const expected = TracingOptions{};
-  EXPECT_EQ(expected.single_line_mode(), actual.single_line_mode());
-  EXPECT_EQ(expected.use_short_repeated_primitives(),
-            actual.use_short_repeated_primitives());
-  EXPECT_EQ(expected.truncate_string_field_longer_than(),
-            actual.truncate_string_field_longer_than());
-}
-
-TEST(PopulateGrpcOptions, DefaultTracingOptionsWithValue) {
+TEST(PopulateGrpcOptions, TracingOptions) {
   ScopedEnvironment env("GOOGLE_CLOUD_CPP_TRACING_OPTIONS",
-                        "single_line_mode=on"
-                        ",use_short_repeated_primitives=ON"
-                        ",truncate_string_field_longer_than=42");
-  auto const actual = internal::DefaultTracingOptions();
-  EXPECT_TRUE(actual.single_line_mode());
-  EXPECT_TRUE(actual.use_short_repeated_primitives());
-  EXPECT_EQ(42, actual.truncate_string_field_longer_than());
+                        "truncate_string_field_longer_than=42");
+  auto actual = PopulateGrpcOptions(Options{}, "");
+  auto tracing = actual.get<GrpcTracingOptionsOption>();
+  EXPECT_EQ(tracing.truncate_string_field_longer_than(), 42);
 }
 
 }  // namespace
